[ts-sdk-v2] Authentication key testing and message improvements
